feat: Working habit creation endpoint
There is a need to implement a draft for User related classes to fully
implement Habit and goal
