+++ app/src/main/java/com/dreamtracker/app/controller/GoalController.java
@@ -1,0 +1,8 @@
+package com.dreamtracker.app.controller;
+
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class GoalController {
+}
