+++ app/src/main/java/com/dreamtracker/app/devTools/CurrentUserProvider.java
@@ -1,0 +1,7 @@
+package com.dreamtracker.app.devTools;
+
+import java.util.UUID;
+
+public interface CurrentUserProvider {
+    UUID getCurrentUser();
+}
