+++ app/src/main/java/com/dreamtracker/app/devTools/DatabaseSeeder.java
@@ -1,0 +1,4 @@
+package com.dreamtracker.app.devTools;
+
+public class DatabaseSeeder {
+}
