+++ app/src/main/java/com/dreamtracker/app/devTools/User.java
@@ -1,0 +1,6 @@
+package com.dreamtracker.app.devTools;
+
+import java.util.UUID;
+
+public record User(UUID id) {
+}
