+++ app/src/main/java/com/dreamtracker/app/entity/Goal.java
@@ -1,0 +1,4 @@
+package com.dreamtracker.app.entity;
+
+public class Goal {
+}
