+++ app/src/main/java/com/dreamtracker/app/repository/GoalRepository.java
@@ -1,0 +1,7 @@
+package com.dreamtracker.app.repository;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface GoalRepository {
+}
