+++ app/src/main/java/com/dreamtracker/app/repository/HabitTrackRepository.java
@@ -1,0 +1,11 @@
+package com.dreamtracker.app.repository;
+
+import com.dreamtracker.app.entity.HabitTrack;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface HabitTrackRepository extends JpaRepository<HabitTrack, UUID> {
+}
