+++ app/src/main/java/com/dreamtracker/app/request/HabitCategoryCreateRequest.java
@@ -1,0 +1,6 @@
+package com.dreamtracker.app.request;
+
+import java.util.UUID;
+
+public record HabitCategoryCreateRequest(UUID id) {
+}
