+++ app/src/main/java/com/dreamtracker/app/request/HabitRequest.java
@@ -1,0 +1,4 @@
+package com.dreamtracker.app.request;
+
+public record HabitRequest(String name, String action, String frequency, String duration, String difficulty) {
+}
