+++ app/src/main/java/com/dreamtracker/app/request/HabitTrackingRequest.java
@@ -1,0 +1,6 @@
+package com.dreamtracker.app.request;
+
+import java.util.UUID;
+
+public record HabitTrackingRequest(UUID habitId, String status) {
+}
