+++ app/src/main/java/com/dreamtracker/app/response/Page.java
@@ -1,0 +1,7 @@
+package com.dreamtracker.app.response;
+
+import java.util.List;
+
+public class Page<T> {
+    private List<T> items;
+}
