+++ app/src/main/java/com/dreamtracker/app/service/GoalService.java
@@ -1,0 +1,4 @@
+package com.dreamtracker.app.service;
+
+public interface GoalService {
+}
