+++ app/src/main/java/com/dreamtracker/app/service/GoalServiceImpl.java
@@ -1,0 +1,8 @@
+package com.dreamtracker.app.service;
+
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class GoalServiceImpl {
+}
