+++ app/src/main/java/com/dreamtracker/app/service/HabitTrackService.java
@@ -1,0 +1,13 @@
+package com.dreamtracker.app.service;
+
+import com.dreamtracker.app.entity.Habit;
+import com.dreamtracker.app.entity.HabitTrack;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface HabitTrackService {
+    Optional<HabitTrack> save(HabitTrack habitTrack);
+    Optional<Habit>findHabitTrackById(UUID id);
+    void deleteById(UUID id);
+}
