+++ app/src/main/java/com/dreamtracker/app/utils/HabitDifficulty.java
@@ -1,0 +1,5 @@
+package com.dreamtracker.app.utils;
+
+public enum HabitDifficulty {
+    EASY,MEDIUM,HARD
+}
