+++ app/src/main/java/com/dreamtracker/app/utils/HabitFrequency.java
@@ -1,0 +1,5 @@
+package com.dreamtracker.app.utils;
+
+public enum HabitFrequency {
+    DAILY,WEEKLY,MONTHLY
+}
