+++ app/src/main/java/com/dreamtracker/app/utils/HabitStatus.java
@@ -1,0 +1,5 @@
+package com.dreamtracker.app.utils;
+
+public enum HabitStatus {
+    ACTIVE,INACTIVE
+}
