+++ app/src/main/java/com/dreamtracker/app/utils/HabitTrackStatus.java
@@ -1,0 +1,5 @@
+package com.dreamtracker.app.utils;
+
+public enum HabitTrackStatus {
+    DONE,UNDONE
+}
