--- app/src/main/java/com/dreamtracker/controller/GoalController.java
@@ -1,4 +1,0 @@
-package com.dreamtracker.controller;
-
-public class GoalController {
-}
