--- app/src/main/java/com/dreamtracker/controller/HabitController.java
@@ -1,4 +1,0 @@
-package com.dreamtracker.controller;
-
-public class HabitController {
-}
