--- app/src/main/java/com/dreamtracker/devTools/CurrentUserProvider.java
@@ -1,7 +1,0 @@
-package com.dreamtracker.devTools;
-
-import java.util.UUID;
-
-public interface CurrentUserProvider {
-    UUID getCurrentUser();
-}
