--- app/src/main/java/com/dreamtracker/devTools/DatabaseSeeder.java
@@ -1,4 +1,0 @@
-package com.dreamtracker.devTools;
-
-public class DatabaseSeeder {
-}
