--- app/src/main/java/com/dreamtracker/devTools/User.java
@@ -1,6 +1,0 @@
-package com.dreamtracker.devTools;
-
-import java.util.UUID;
-
-public record User(UUID id) {
-}
