--- app/src/main/java/com/dreamtracker/entity/Goal.java
@@ -1,4 +1,0 @@
-package com.dreamtracker.entity;
-
-public class Goal {
-}
