--- app/src/main/java/com/dreamtracker/repository/GoalRepository.java
@@ -1,7 +1,0 @@
-package com.dreamtracker.repository;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface GoalRepository {
-}
