--- app/src/main/java/com/dreamtracker/repository/HabitRepository.java
@@ -1,11 +1,0 @@
-package com.dreamtracker.repository;
-
-import com.dreamtracker.entity.Habit;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface HabitRepository extends JpaRepository<Habit, UUID> {
-}
