--- app/src/main/java/com/dreamtracker/repository/HabitTrackRepository.java
@@ -1,11 +1,0 @@
-package com.dreamtracker.repository;
-
-import com.dreamtracker.entity.HabitTrack;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface HabitTrackRepository extends JpaRepository<HabitTrack, UUID> {
-}
