--- app/src/main/java/com/dreamtracker/request/HabitCategoryCreateRequest.java
@@ -1,6 +1,0 @@
-package com.dreamtracker.request;
-
-import java.util.UUID;
-
-public record HabitCategoryCreateRequest(UUID id) {
-}
