--- app/src/main/java/com/dreamtracker/request/HabitRequest.java
@@ -1,4 +1,0 @@
-package com.dreamtracker.request;
-
-public record HabitRequest(String name, String action, String frequency, String duration, String difficulty) {
-}
