--- app/src/main/java/com/dreamtracker/request/HabitTrackingRequest.java
@@ -1,6 +1,0 @@
-package com.dreamtracker.request;
-
-import java.util.UUID;
-
-public record HabitTrackingRequest(UUID habitId, String status) {
-}
