--- app/src/main/java/com/dreamtracker/response/HabitResponse.java
@@ -1,6 +1,0 @@
-package com.dreamtracker.response;
-
-import java.util.UUID;
-
-public record HabitResponse(UUID id, String name, String action, String duration, String difficulty, String status) {
-}
