--- app/src/main/java/com/dreamtracker/response/Page.java
@@ -1,7 +1,0 @@
-package com.dreamtracker.response;
-
-import java.util.List;
-
-public class Page<T> {
-    private List<T> items;
-}
