--- app/src/main/java/com/dreamtracker/service/GoalService.java
@@ -1,4 +1,0 @@
-package com.dreamtracker.service;
-
-public interface GoalService {
-}
