--- app/src/main/java/com/dreamtracker/service/GoalServiceImpl.java
@@ -1,8 +1,0 @@
-package com.dreamtracker.service;
-
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class GoalServiceImpl {
-}
