--- app/src/main/java/com/dreamtracker/service/HabitTrackService.java
@@ -1,13 +1,0 @@
-package com.dreamtracker.service;
-
-import com.dreamtracker.entity.Habit;
-import com.dreamtracker.entity.HabitTrack;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface HabitTrackService {
-    Optional<HabitTrack> save(HabitTrack habitTrack);
-    Optional<Habit>findHabitTrackById(UUID id);
-    void deleteById(UUID id);
-}
