--- app/src/main/java/com/dreamtracker/utils/HabitDifficulty.java
@@ -1,5 +1,0 @@
-package com.dreamtracker.utils;
-
-public enum HabitDifficulty {
-    EASY,MEDIUM,HARD
-}
