--- app/src/main/java/com/dreamtracker/utils/HabitFrequency.java
@@ -1,5 +1,0 @@
-package com.dreamtracker.utils;
-
-public enum HabitFrequency {
-    DAILY,WEEKLY,MONTHLY
-}
