--- app/src/main/java/com/dreamtracker/utils/HabitStatus.java
@@ -1,5 +1,0 @@
-package com.dreamtracker.utils;
-
-public enum HabitStatus {
-    ACTIVE,INACTIVE
-}
