--- app/src/main/java/com/dreamtracker/utils/HabitTrackStatus.java
@@ -1,5 +1,0 @@
-package com.dreamtracker.utils;
-
-public enum HabitTrackStatus {
-    DONE,UNDONE
-}
