+++ app/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=app
+spring.application.name=dream-tracker
 spring.datasource.url=jdbc:postgresql://localhost:5232/dream-tracker
 spring.datasource.username=user
 spring.datasource.password=password
